refactor common extraction logic
