+++ src/common/mod.rs
@@ -1,1 +1,2 @@
+pub(crate) mod prebuilt;
 pub(crate) mod union;
